% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/sampling.R
\name{betasUnder}
\alias{betasUnder}
\title{betasUnder}
\usage{
betasUnder(y, positive = 1, N = 10, method = "perc")
}
\arguments{
\item{y}{response variable}

\item{positive}{value of the positive (minority) class}

\item{N}{number of values of beta to return}

\item{method}{method to compute beta: perc or prob}
}
\value{
values of beta for a give response variable
}
\description{
Defines the possible levels of undersampling given the class proporiton
}
\examples{
y <- rep(c(1, 0, 1, 0, 0, 0), 100)
betasUnder(y, positive=1, N = 10, method="perc")
}

