% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/warping.R
\name{brierScore}
\alias{brierScore}
\title{Brier Score}
\usage{
brierScore(phat, truth, positive = 1)
}
\arguments{
\item{phat}{probability of a instance to belogn to the class positive}

\item{truth}{class of the instance}

\item{positive}{value of the positive (minority) class}
}
\value{
Brier Score statistics
}
\description{
Function used to compute Brier Score, a measure of probability calibration.
}
\examples{
truth <- c(1,1,1,0,0,0,1,0,1,0)
phat <- c(0.9,0.8,0.4,0.5,0.3,0.2,0.8,0.3,0.8,0.3)
brierScore(phat, truth, 1)
}

