% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/warping.R
\name{BS}
\alias{BS}
\title{warpingWrapper: Wrapper to warpingUnder function
#
#}
\usage{
BS(phat, truth, positive)
}
\arguments{
\item{phat}{probability of a instance to belogn to the class positive}

\item{truth}{class of the instance}

\item{positive}{value of the positive (minority) class}

\item{algo}{classification algorithm supported by mlr package
#}

\item{dataset}{dataset used for the analysis
#}

\item{nCV}{number of repetation of the Cross Validation (CV)
#}

\item{B}{number of models to create for each fold of the CV
#}

\item{nFold}{number of folds of the CV
#}

\item{ncore}{number of cores to use in multicore computation
#}

\item{...}{extra parameters passed to mlr train function
#
#}
}
\value{
The function returns a list:
#\item{results}{results of undersampling}
#\item{probs}{probabilities}
#\item{mres}{mres}
#\item{meltResALL}{meltResALL}
#
#

Brier Score statistics
}
\description{
#Calls the function \code{\link{warpingUnder}} using a classisification algorithm and an unbalanced dataset
#
#
}
\details{
#
}
\examples{
#library(warping)
#warpingWrapper("randomForest", "pima", nCV=50, nFold=3)
#
#
#
}

