% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/warping.R
\name{warpingWrapper}
\alias{warpingWrapper}
\title{warpingWrapper: Wrapper to warpingUnder function}
\usage{
warpingWrapper(algo = "randomForest", dataset = "pima", nCV = 10, B = 1,
  nFold = 100, ncore = 1, ...)
}
\arguments{
\item{algo}{classification algorithm supported by mlr package}

\item{dataset}{dataset used for the analysis}

\item{nCV}{number of repetation of the Cross Validation (CV)}

\item{B}{number of models to create for each fold of the CV}

\item{nFold}{number of folds of the CV}

\item{ncore}{number of cores to use in multicore computation}

\item{...}{extra parameters passed to mlr train function}
}
\value{
None
}
\description{
Calls the function \code{\link{warpingUnder}} using a classisification algorithm and an unbalanced dataset
}
\examples{
warpingWrapper("randomForest", "pima", nCV=50, nFold=3)
}

